# core/api/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "game/v1/engine.proto",
    "description": "Package game.v1 is the Core service boundary — the central product surface.\nCore serves this contract over both gRPC and REST: the google.api.http\nannotations below drive a grpc-gateway that exposes every RPC as JSON/HTTP\nunder /api/v1, wrapped in the uniform {code, message, trace_id, data}\nenvelope. Core is auth-agnostic — it trusts the caller (a BFF the developer\nbuilds, see examples/) to authenticate and to pass the tenant/merchant\nisolation keys (tenant_id + merchant_id); it only validates and operates on\nthe business object. Errors map to gRPC status codes (and to the envelope at\nthe gateway). Versioned by package; a breaking change becomes game.v2.",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/games/{game_id}/milestones": {
      "get": {
        "operationId": "WalletService_GetMilestones",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1GetMilestonesResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "game_id",
            "description": "Game whose milestones to read (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenant_id",
            "description": "Tenant isolation key (required).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "merchant_id",
            "description": "Merchant isolation key within the tenant.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "player_id",
            "description": "Player whose progress to read.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletService"
        ]
      }
    },
...
    "/api/v1/games/{game_id}/redeem": {
      "post": {
        "operationId": "WalletService_Redeem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1RedeemResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "game_id",
            "description": "Game the milestone belongs to (from the path).",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletServiceRedeemBody"
            }
          }
        ],
        "tags": [
          "WalletService"
        ]
      }
    },
...
    "/api/v1/wallet": {
      "get": {
        "operationId": "WalletService_GetWallet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1GetWalletResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "tenant_id",
            "description": "Tenant isolation key (required).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "merchant_id",
            "description": "Merchant isolation key within the tenant.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "scope_key",
            "description": "Wallet scope key (e.g. campaign id).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "player_id",
            "description": "Player whose wallet to read.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletService"
        ]
      }
    },
    "/api/v1/wallet/ledger": {
      "get": {
        "operationId": "WalletService_GetLedger",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int32"
                },
                "message": {
                  "type": "string"
                },
                "trace_id": {
                  "type": "string"
                },
                "data": {
                  "$ref": "#/definitions/v1GetLedgerResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/apiErrorEnvelope"
            }
          }
        },
        "parameters": [
          {
            "name": "tenant_id",
            "description": "Tenant isolation key (required).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "merchant_id",
            "description": "Merchant isolation key within the tenant.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "scope_key",
            "description": "Wallet scope key (e.g. campaign id).",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "player_id",
            "description": "Player whose ledger to read.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "description": "Max entries to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "cursor",
            "description": "Opaque pagination cursor from a previous response.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletService"
        ]
      }
    }
  },
  "definitions": {
...
    "WalletServiceRedeemBody": {
      "type": "object",
      "properties": {
        "tenant_id": {
          "type": "string",
          "description": "Tenant isolation key (required)."
        },
        "merchant_id": {
          "type": "string",
          "description": "Merchant isolation key within the tenant."
        },
        "milestone_id": {
          "type": "string",
          "description": "Milestone to redeem."
        },
        "player_id": {
          "type": "string",
          "description": "Player redeeming."
        }
      },
      "description": "RedeemRequest spends points to claim a milestone reward."
    },
...
    "v1GetLedgerResponse": {
      "type": "object",
      "properties": {
        "entries": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1LedgerEntry"
          },
          "description": "Page of ledger entries (newest first)."
        },
        "next_cursor": {
          "type": "string",
          "description": "Cursor for the next page (\"\" when exhausted)."
        }
      }
    },
...
    "v1GetMilestonesResponse": {
      "type": "object",
      "properties": {
        "currency": {
          "type": "string",
          "description": "Currency the milestones are denominated in."
        },
        "mode": {
          "$ref": "#/definitions/v1MilestoneMode",
          "description": "How the milestone set behaves."
        },
        "balance": {
          "type": "string",
          "format": "int64",
          "description": "Player's current balance in the currency."
        },
        "milestones": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1MilestoneState"
          },
          "description": "Milestones with the player's progress."
        }
      }
    },
...
    "v1GetWalletResponse": {
      "type": "object",
      "properties": {
        "balances": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          },
          "description": "Balances keyed by currency."
        }
      }
    },
...
    "v1LedgerEntry": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Stable unique identifier."
        },
        "currency": {
          "type": "string",
          "description": "Currency code the entry is in."
        },
        "amount": {
          "type": "string",
          "format": "int64",
          "description": "Signed amount: + earn, - spend."
        },
        "reason": {
          "$ref": "#/definitions/v1LedgerReason",
          "description": "Why the entry was written."
        },
        "ref_id": {
          "type": "string",
          "description": "Reference id (play, redemption, …)."
        },
        "created_at": {
          "type": "integer",
          "format": "int64",
          "description": "Entry time (unix seconds)."
        }
      },
      "description": "LedgerEntry is one signed movement in a player's points wallet."
    },
    "v1LedgerReason": {
      "type": "string",
      "enum": [
        "LEDGER_REASON_UNSPECIFIED",
        "LEDGER_REASON_PLAY",
        "LEDGER_REASON_REDEEM",
        "LEDGER_REASON_ADJUST"
      ],
      "default": "LEDGER_REASON_UNSPECIFIED",
      "description": "LedgerReason is why a wallet ledger entry was written."
    },
...
    "v1MilestoneMode": {
      "type": "string",
      "enum": [
        "MILESTONE_MODE_UNSPECIFIED",
        "MILESTONE_MODE_CUMULATIVE_UNLOCK",
        "MILESTONE_MODE_SPEND_EXCHANGE"
      ],
      "default": "MILESTONE_MODE_UNSPECIFIED",
      "description": "MilestoneMode is how a milestone set behaves."
    },
    "v1MilestoneState": {
      "type": "object",
      "properties": {
        "milestone_id": {
          "type": "string",
          "description": "Milestone id."
        },
        "threshold": {
          "type": "string",
          "format": "int64",
          "description": "Points threshold to unlock."
        },
        "prize_id": {
          "type": "string",
          "description": "Prize granted at the threshold."
        },
        "status": {
          "$ref": "#/definitions/v1MilestoneStatus",
          "description": "Progress state for the player."
        },
        "progress": {
          "type": "string",
          "format": "int64",
          "description": "Current progress toward the threshold."
        },
        "remaining": {
          "type": "string",
          "format": "int64",
          "description": "Points remaining to the threshold."
        }
      },
      "description": "MilestoneState is a milestone plus the caller's progress toward it."
    },
    "v1MilestoneStatus": {
      "type": "string",
      "enum": [
        "MILESTONE_STATUS_UNSPECIFIED",
        "MILESTONE_STATUS_LOCKED",
        "MILESTONE_STATUS_UNLOCKED",
        "MILESTONE_STATUS_GRANTED"
      ],
      "default": "MILESTONE_STATUS_UNSPECIFIED",
      "description": "MilestoneStatus is a milestone's progress state for a player."
    },
...
    "v1RedeemResponse": {
      "type": "object",
      "properties": {
        "redeemed": {
          "type": "boolean",
          "description": "Whether the redemption succeeded."
        },
        "mode": {
          "$ref": "#/definitions/v1MilestoneMode",
          "description": "Milestone mode that governed the redemption."
        },
        "spent": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          },
          "description": "Amounts spent, keyed by currency."
        },
        "reward": {
          "$ref": "#/definitions/v1Reward",
          "description": "Reward granted, if any."
        },
        "balances": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "int64"
          },
          "description": "Resulting balances, keyed by currency."
        }
      }
    },
...
    "v1Reward": {
      "type": "object",
      "properties": {
        "prize_id": {
          "type": "string",
          "description": "Prize that was awarded."
        },
        "name": {
          "type": "string",
          "description": "Prize name for display."
        },
        "image": {
          "type": "string",
          "description": "Prize image URL for display."
        },
        "type": {
          "type": "string",
          "description": "Prize type."
        },
        "quantity": {
          "type": "string",
          "format": "int64",
          "description": "Number of units awarded."
        },
        "value": {
          "type": "string",
          "format": "int64",
          "description": "Nominal value for display."
        },
        "reward_id": {
          "type": "string",
          "description": "Durable reward record id (used to claim)."
        },
        "code": {
          "type": "string",
          "description": "Assigned voucher code, if any."
        }
      },
      "description": "Reward is a single awarded item in a play result."
    },
...
    "apiFieldViolation": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "apiErrorInfo": {
      "type": "object",
      "description": "data.error payload (mirrors google.rpc.ErrorInfo).",
      "properties": {
        "status": {
          "type": "string",
          "description": "canonical gRPC status name"
        },
        "reason": {
          "type": "string",
          "description": "stable machine reason (see docs/ERRORS.md)"
        },
        "domain": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "field_violations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiFieldViolation"
          }
        }
      }
    },
    "apiErrorData": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/apiErrorInfo"
        }
      }
    },
    "apiErrorEnvelope": {
      "type": "object",
      "description": "Uniform error envelope.",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "trace_id": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/apiErrorData"
        }
      }
    }
  }
}